Fix some of the lints triggered by future versions of Clippy
